+++ test/k6/tag.js
@@ -1,50 +1,13 @@
-import http from "k6/http";
-import {check, group, sleep, fail} from 'k6';
+import { sleep} from 'k6';
+import {
+  post_gql,
+  setup as setup_helper
+} from './helpers.js';
 
 // export let options = {
 //     vus: 10,
 //     duration: '1s',
 // };
-
-const USERPHONE = `917834811114`;
-const PASSWORD = 'secret1234';
-const BASE_URL = 'http://glific.test:4000';
-
-export function setup() {
-    // register a new user and authenticate via a Bearer token.
-    let loginRes = http.post(`${BASE_URL}/api/v1/session`, {
-        "user[phone]": USERPHONE,
-        "user[password]": PASSWORD
-    });
-
-    let access_token = loginRes.json('data').access_token;
-    check(access_token, { 'logged in successfully': () => access_token !== '', });
-
-    return access_token;
-}
-
-function post_gql(query, access_token) {
-    let headers = {
-        'Authorization': access_token,
-        "Content-Type": "application/json"
-    };
-
-    let res = http.post(`${BASE_URL}/api`,
-                        JSON.stringify({ query: query }),
-                        {headers: headers}
-                       );
-
-    check(res, {
-        'is status 200': (r) => r.status === 200,
-    });
-
-    // console.log(JSON.stringify(res.json('data')));
-    if (res.status !== 200) {
-        console.log(JSON.stringify(res.body));
-    };
-
-    return res.json('data');
-}
 
 function tags_query() {
     return `
@@ -97,6 +60,11 @@ function tags_delete_query(id) {
       }
     `;
 }
+
+export function setup() {
+  return setup_helper()
+}
+
 export default function(access_token) {
     post_gql(tags_query(), access_token);
     sleep(1)
